+++ src/pages/visualization.py
@@ -3,7 +3,3 @@ from st_pages import show_pages_from_config, add_page_title
 # Either this or add_indentation() MUST be called on each page in your
 # app to add indendation in the sidebar
 add_page_title()
-
-
-
-
